Add lazy loading and code splitting for modals
